detection-system
enemymovement
function that moves the enemy to a certain point on the map.
add array of points.
function that moves the enemy to the point after the enemy have reached the first point.
check if player reached a point then add one more point.
Enemies
ray cast that destroys the player if detected.