Raycast to the turret
Added Raycast to the turret and a laser effect so that the player knows what area they must avoid.
Raycast debugs perfectly fine.

To do:
Make another raycast just for the turret deployement and keep another raycast to shoot so if the player goes closer to the turret it shoots.