Combining Team's work before build of the week
~Made the pause screen into a prefab and overid keypad prefab to transfer to main game scene.

~organised player 2 movement script a bit

~Pausing the game puts to time scale 0, if you got to main menu and come back the game is stuck on timescale 0, made it so it goes back to time scale 1

~Reduced the size of combat player's collidor, and reduced doorway size a bit so he doesn't get stuck in the doorway

~Tried figuring out how soham's attack script works and just increased the attack radius as a temporary fix,
Will make it so collidors move based on attack sprite animation for attack to create accuracy.

~Added visual get hit for combat player

~Fixed player movement stopped when keypad being used, made it so the movement comes back after keypad is closed with the button rather than just pressing E
Stuff is a bit messy,

The close button when pressed goes to the keypad functiinality script where as the stopping of player movement is in interactiveKeypad script
which is only used when E is pressed on the keypad hence not enabling the movement again when close button is pressed.

which took me a while to figure out and finally got the simple fix which was just to place the same code to disable player script in the keypad funtionality for the close button, cross referencing.

~after making build found that the movement isn't even given back after putting in the right code either
so fixing that now too

Added the same line under the correct answer in keypad section

~added the same keypad sound to the button before it for water puzzle, by adding audio source and clip to the combat player