Health bar
Ran into an issue due to the last pull. The issue was that the combat player in my scene could not jump because insteam on Oncollisionstay we added a raycast to the player. Therefore, to keep the work going I made a copy of the main scene and worked in it for my health bar mechanic.

Made two new scripts called "Player_1_health" and "Player_2_health".
The health decreases when the player is colliding with the enemy raycast