Updated Main Menu
Added a background in the "Main Menu" scene. For this I used the environment prefab and added it. The prefab had rain effect so that the menu looks alive.

Also added custon buttons.
Added a new button called "Controls" when clicked, it opens a panel where the player can learn controls. Also added a back button in the panel to deactivate the panel and go back to menu.
Imported a new font "alagard" and used it on the button texts